Orbit star function implementation.

+++ Entities/CelestialBody.cs
@@ -1,7 +1,4 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+
 
 namespace space_exploration
 {
@@ -9,7 +6,7 @@ namespace space_exploration
     {
 
         public string Name {get;set;} 
-        public double Size {get;set;}
+        public float Size {get;set;}
         public int Age {get;set;}
         public float Mass {get;set;}
         public float Gravity {get;set;}
